Fix proprietary save file read buffer typing and verify tests
Agent-Logs-Url: https://github.com/KoshkiKode/cordite/sessions/0403dd27-45f7-426e-bea3-c417f3cc6da3

+++ src/Systems/Persistence/SaveManager.cs
@@ -154,7 +154,7 @@ public sealed partial class SaveManager : Node
                 return null;
             }
 
-            byte[] payload = file.GetBuffer(file.GetLength());
+            byte[] payload = file.GetBuffer((long)file.GetLength());
             SaveGameData? data = SaveFileCodec.Decode(payload, SaveJsonOptions);
 
             if (data != null)
@@ -256,7 +256,7 @@ public sealed partial class SaveManager : Node
                     using var file = FileAccess.Open(filePath, FileAccess.ModeFlags.Read);
                     if (file is not null)
                     {
-                        byte[] payload = file.GetBuffer(file.GetLength());
+                        byte[] payload = file.GetBuffer((long)file.GetLength());
                         SaveGameData? data = SaveFileCodec.Decode(payload, SaveJsonOptions);
 
                         if (data != null)
